Add basic VFX unit debug window
On a private test branch, there are extensive crashes in the VFX subsystem
when saving and loading states. Rather than attempt to merge that branch
with crashes or attempt to build debug tooling on a private branch that's
already quite large, I broke out the VFX tooling into a separate commit.