Run memento operations performed by the overlay in the main loop
Repeatedly and quickly clicking on any of the memento operations in the
overlay seems to eventually crash the game. It may be that there are some
edge cases around whether memento operations occur during the update step
(like the existing training mode F9/F10 buttons, which does not crash), or
in the render step (where the ImGui overlay code runs). To avoid those
hypothetical cases, the overlay now sets a value in
sf4e::Game::Battle::System when the user clicks the button, and a detour
inside the update step executes the requested operation if that value is
set. This necessitated moving the "manual" restore to a helper method that
wasn't in the overlay.

This does seem to fix the crashes- repeatedly and quickly clicking on
the new "Request" options don't seem to crash, no matter how long or
fast I click. However, just in case it proves useful later, the UI still
has support for "immdediate" operations.

While I was here, I reordered the manipulation of states in the
"manual" methods, to make sure we mirror the existing implementation.
There may be ordering constraints in the state saves and loads.