Track memento sizes and log when they vary
SF4's mementos are reinitialized on every save attempt, for various
reasons. If sf4e is to implement multiple save states using SF4's internal
mementos, we need to prevent reinitialization of all mementos on a save,
and either only reinitialize the saved memento, or reinitialize none of
them and allow all mementos to hold all states. The latter is much
easier, and is what most games would do, but SF4's memento sizes are
currently dynamic- and all memento sizes are allocated up front. This
means that if a particle doesn't exist the first time the state is saved
to a memento, and we prevent reinitialization, there's _never_ memory to
save that particle.

By tracking which objects vary in size, we can eventually fix the size
to be the maximum size identified.