Add spdlog
Logging in ImGui disappears immediately once the process crashes, and
can't be presented to a developer unless ImGui can render. Logging to a
file, and in particular using spdlog for various conveniences around
that logging, should help developers debug crashes as well as provide a
way to exchange logfiles that aren't screenshots.