CharaSelect: Fix PlayerCondition starting offset, include flags
The memory at the address I had originally marked as the start of a
PlayerCondition pair is _actually_ the set of flags used by the
character select screen to enable or disable various features, and the
PlayerCondition pair starts 4 bytes later.. Since the size of the
PlayerCondition struct was correct, array access still worked, which
is why the overlay worked for both players. That being said, it still
seems like a footgun to have this hanging around, and sf4e will
likely need to be able to access and manipulate the flags anyway.

While I was here, I included an initial analysis of flags as an enum.