Hook SF4's internal std::string methods
While SF4 seems to use std::string internally and std::string, pointers
can actually be used to read data, they can't naively be used to _write_
data. It's unclear if this is because SF4 uses a custom allocator for
strings, because the DLL is linking against a different std::string
implementation that could be solved with a different msvcrt, or another
unidentified reason. Regardless, we can hack around all of these by
de-templatizing" the allocator and locating the allocator methods as
if they were method pointers of a string.