Mementos: Add logging when unsafe initialization skipping is wrong
When skipping initialization, memento key size debug logging can't
compare against the old size of the key because the key's size isn't
changing in the first place. Instead, create and initialize a new key,
then compare the existing single-initialization key size against the
fresh temporary key's size.

This also fixes a possible subtle bug in initialization skipping.
sf4e used to only check that the size and mementoable matched, but it's
hypothetically possible for the size and mementoable to both be _zero_,
or otherwise for a constructed but uninitialized key to match the
initialization args. Since it's guaranteed that a key's metadata
pointer will be set after one initialization, skipping initialization
if the metadata pointer exists guarantees that we don't try to skip
initialization on an uninitialized key.