Main menu: create shortcut to Versus interstitial screen
This is an alternative way to enter versus mode from the main menu
overlay, that allows the user to skip the character and stage selection.
Consequently, the overlay exposes UI elements to select the necessary
values identified during character and stage selection.

When this method is used, both characters are bound to the same
controller. This is wildly incorrect, but is a useful stepping stone for
future extension.