Support simulating without HUD updates
SF4's built-in memento system doesn't properly track many HUD elements-
it doesn't track the scale or timing of the "FIGHT" sprite at the beginning
of a round, for example. Consequently, it's impossible to correctly
save, load, and resimulate the HUD.

Since the HUD doesn't have any gameplay consequences and any incorrect
HUD state will be fixed the next time a HUD message is sent on an
update, we can actually just avoid resimulation of the HUD entirely and
only simulate the HUD on user-visible frames. Each frame is only visible
to the user once, ensuring that animation timing will stay consistent.