Battle System: Add battle-exit control
When the battle flow enters BTL_OVER (flow 11), the battle system is
updated to be in the "is leaving battle" state. The wrapping event (ex.
VsBattle) checks the battle state and the exit type before determining
what to do. In some conditions, the wrapping event will display the
results menu, and in some conditions it will use the event flow to
navigate either to character select or the main menu.

This adds an option to force VsBattle to enter a "terminal" state, to
go back to the main menu, which exercises most of the above.