Battle system: Demonstrate skipping from the battle-start flow
The battle-start flow (flow 0xa) is the initial flow when entering a
battle. The flow handler for battle-start is dead simple it is
responsible for determining where to go next based on the current game
type- for most matches, it goes to character demos, and for training and
trials, it cuts right to the round. Replacing the destination of that
handler provides an easy way to allow skipping character demos when first
entering a versus match, for example.

However, the initial flow can't be replaced directly when entering a
battle. Attempting to set the initial flow to anything other
than battle-start _on the first battle that the system is used_ (such as
by calling the system's `Restart()` method) results in segfaults, likely
due to global state not being set up. See 0x5dc150, which actually sets
the initial flow and performs initial setup.