GGPO: Use P1's controls on both server and client via setting IsOnlineBattle
The proof of concept relied on setting P1's controller to the first
keyboard and P2's controller to the second keyboard in order to work
around the need for a controller selection UI. However, even if sf4e
allows for selecting controllers, it also needs to work with _only one_
controller, because it's difficult to identify a connected-but-unused
controller slot to represent the remote player.

When offline, SF4 pauses when the local device used by a player is
disconnected. When online, there is by definition only one local device,
so the disconnect check must be skipped. IsOnlineBattle is a check that
exists in SF4's delay-based online infrastrure, and sf4e can force it to
be true for GGPO matches to prevent the pause check from halting the
game. This means we only need one active device on each side, and
consequently that each side can use the "P1" slot.