GGPO: Fix throw crash by supporting variable-length memento lists
With the original goal of a fixed-size save state declarable in a
header, it was important that we know in advance how many memento keys
could possibly be in a SaveState. However, effect creation (such as
throws creating the ground shockwave effect can strangely result in
new memento keys being initialized. The effect objects themselves are
preallocated, so it _should_ still be safe to copy these
late-initialized memento keys back into place.

Instead of statically searching for a new upper bound or adding debug
tools to identify a high-water mark, just use vectors with .reserve()
to minimize memory-allocation overhead. Warnings in the logfile are
still useful, but they should no longer alert the user and halt the
process like a message-box would.