GGPO / Loading states: Fix super-combo meter "popping" to screen center
Demonstrated here: https://mastodon.social/@lulolwen/110733620472954122

To easily reproduce:
1. Be between 50% and 75% meter
2. Save state
3. Use one EX move
4. Disable HUD updates
5. Load state
6. Re-enable HUD updates

Dimps::Game::Battle::Hud::Cockpit::Sc manages at least 4 sprites, one
for each "part" of the SC meter (each 25% division). Its load method
@ 0x5792c0 does more than just copy state out of the memento and into
the Sc control, ut it attempts to re-stage the SC "part" sprites.
However, it doesn't actually set the position of the sprites- it only
queues them for rendering. Dimps::Game::Battle::Hud::Cockpit::View::Update
@ 0x5764b0 is responsible for setting the position to reflect the user's
HUD customization. This is impossible to notice in unmodified SF4
because all loads occur at the top of an update, and then the HUD
updates run and reconcile the position, but is obvious when running in
GGPO.

`ggpo_advance_frame()` not only saves a new state, but also frees old
ones that are no longer necessary. Freeing a state in sf4e's GGPO integration
right now actually requires loading the victim state before freeing it,
since the mementos use pointers to reference instances of mementoable
objects, and the instance is called to free the memento. Since
`ggpo_advance_frame()` is called after _all_ updates (including the HUD)
but before rendering, the implicit load inside a free results in
re-staging the Sc parts in the center of the screen.

Just make sure to update when the View is restored from a memento. The
update when the memento is restored won't affect animations, due to
forcing the animation timescale to animate 0 frames, despite calling
`Update()`.