GGPO save states: Place "loose" members into a struct
This is a convenient step in checksumming and state comparison- the
single object can be memcpy'd and memcmp'd, rather than needing to
compare members of the SaveState directly.

Since I wanted the new GlobalData struct to be a proper aggregate, that
also required moving initialization to the header rather than a
user-defined constructor.