GGPO + skip-to-versus: Fix softlock with character of invalid edition
The "edition" fields in both the network-test menu and the
skip-to-versus menu allow the user to specify invalid combinations of
characters and editions, like "SSF4 Yun". If an invalid combination is
attempted, SF4 correctly advances to the versus interstitial screen, and
then softlocks because it's waiting for a file that doesn't exist to be
loaded (and obviously, the file is never loaded). Compounding this is
that if an edition outside the ones available via the game's normal
selection is chosen, the versus interstitial may lie to you- for
example, edition "0" appears to be SSF4 1.0, which will appear on the
loading screen as USF4.

Instead, leverage the game's existing valid-edition-per-character
list, and turn the integer input into a combo box that ensures only
valid choices are selected. This implicitly forces a valid edition when
the character changes as well.