GGPO / mementos: Persist/load Battle::Network::Unit state
Fixes https://gitlab.com/adanducci/sf4e/-/issues/1 .

All inputs go into and out of the network unit, but frames must be in
the network unit for at least one frame. Because the single queued frame
of input in Network::Unit isn't saved and loaded with the rest of the
CommandImpl, save states created exactly 1f after an input miss the
input.

This is very much a shotgun solution- we technically only need to save
and load a single value in order to fix this specific state desync- but
I don't understand the Network::Unit's input queueing well enough to say
that for sure, and it doesn't appear to actually have pointers in it so
this should be safe.