GGPO / mementos: Fix duplicate KO/round start announces
Fixes https://gitlab.com/adanducci/sf4e/-/issues/2 .

Just more state that's missing from SF4's save states. This isn't
relevant in the vanilla SF4 engine because it's impossible to create
save states during an announcement, but is necessary for GGPO.

Te Battle::System checks if the Announce unit has an active control as
a way of determining if it should set the slowed-down timescale after
a KO and so on- if there's no control, the System believes it has work
to do. The cascade of additional state that we need to save once we
start saving the active control is _long_:

* The active control has sprites, which also need to be saved and loaded
* The sprites have associated actions, which also need to be saved and
  loaded or
* The System has additional data that it uses to track the fade in/fade
  out transitions, as well as flags that determine when the players have
  skipped or when inputs are valid for recording
* The "SYS START" task responsible for resetting the round after the
  reduced timescale has additional state, and the task itself must be
  saved and loaded, which requires _massive_ updates through other
  parts of the system to update references to tasks.

However, we get a lot from this as well- it additionally fixes camera
shake, as well as making it possible to avoid halting the HUD simulation
during `ggpo_advance_frame_callback()`.