GGPO: Stop forcing IsOnlineBattle, and prevent pausing directly
If IsOnlineBattle is true when a match loads, the Network unit creates
tasks that are slightly different than offline. For one, NetProvide is
created, which attempts to interact with the Steam session and replace
inputs with those gathered from the Session subsystem, which we're not
integrated with. For two, if the difference in time between loads is too
long, the NetMain task actually pauses the game to wait for inputs
through the Session subsystem, and since those inputs aren't coming the
game deadlocks.

Instead, just prevent pause directly.