GGPO: Extend session with multiple rounds, spectators, and desync detection
Putting all of these features in a single commit makes the commit hard
to reason about, but easy to develop on- the implementations of each
informed all of the others, primarily due to refactoring the
relationship of SessionClient and SessionServer. The "host" now runs
both a client and a server.

This also includes capturing the active gamepad so that non-keyboards
can be used, actually going back to the main menu when disconnects
happen, and other QOL.