GGPO: Use correct IP address for session host when constructing session
This isn't the "right" way, but it is only like 10LOC.

Originally, the session server sent the IP address of each client out to
all clients so that the clients have all the information they need to
connect to everyone. However, that IP address is based on the IP address
that the server sees- but if the server is running in a listen mode, the
local client uses localhost to connect to the server. This means the
server informs all other clients that the listening client is at IP
127.0.0.1- obviously not a publicly routable IP address.

Instead, just store the server address in each client. When the
listening client connects, its IP address is instead represented by the
empty string, which every client will then replace with the server
address when the GGPO session is initialized. The listen client is
by definition always local, so the IP address is never used by the
listen client. This also works with multiple instances on the same host-
all the instances identify with localhost, so all of them have empty IP
addresses, and all of them use the server address to connect to each
other, which is localhost.

The right way is probably STUN, which we'll probably need at some point,
but using STUN (and therefore canonical IPs for all members in the
session) has downstream consequences on whether or not the server and
listening client actually use the loopback interface.