Allocate console and forward spdlog messages to it
This is a 10-minute solution to providing user feedback when things go
wrong in the GGPO session. Ideally the overlay would have a
flash-message or dialog-box based approach, but that requires quite a
bit more polish. Once that polish is in place, this console should
eventually be disabled via command line arguments or only enabled for
debug builds.