GGPO: Reduce disconnect timeouts to fix ring buffer assertions
It is possible to set a GGPO disconnect timeout that is high enough that
it is impossible for spectators to disconnect without crashing the host.
On a "dirty" disconnection caused by process closure, the host stores
inputs in `udp_proto.h`'s fixed-size 64-frame `_pending_output` buffer,
removing them from the buffer once the inputs are acknowledged by all
connected players and spectators. Once a spectator is considered
disconnected, the host no longer waits for acknowledgements from that
spectator. However, if the disconnection timeout is greater than
`64f / your simrate` and a spectator stops responding for 64 frames or
more, the spectator is still not considered disconnected so a 65th
(or beyond) input must be written to the queue- resulting in an
assertion protecting the host from overflow.

Real-world numbers have a fudge factor beyond the theoretical `64f /
your simrate` limit, but 1000ms is probably a solid line anyway.