Extend random-input tool in GGPO
With the existing tool, since the inputs re-randomize every frame, we
end up with maximum misprediction, leading to almost every frame needing
a rollback. This is really expensive, and not really necessary for local
testing if I just need two automatic instances to run. Instead, use a
configurable randomization speed. This may also help to diagnose
specific timing issues in the future.