Potentially fix wild camera behavior and crashes during rollbacks
(Potentially) fixes #8 .

Task "cores" are probably better named task _pools_: the task objects
inside them are persistent in memory over the lifetime of the pool, and
are added and removed from the "live" list and the "dead" list as tasks
are acquired and released. When the core is created, a large amount of
memory is acquired and each task gets a slice of that "private" data to
use as working-memory. The pointer to that slice is stored in the task
object itself.

The problem is `fTask::RestoreFromAdditionalMemento`. At the moment, it
overwrites almost all of the task object with data from the savestate,
including the private data pointer. That means if a task was originally
stored in the taskcore pool at offset 2 but is being loaded into the
taskcore pool at offset 3, the new task _still points at offset 2's
private data_. In case of the camshake task, accidentally pointing at
a different task's data could result in the camshake code interpreting
that data as an infinite camshake, a zoom-out to infinitity, or a
division by zero- all of which are bad situations.

Just stop overwriting it. The rest of the infrastructure should work the
same.