Delay VsBattle termination to eventually support session draining
Since spectators are not timesynced with players, it's possible for
player clients to enter the terminal state before spectator clients do.
There are two possible approaches here:

* Allow players to exit the battle immediately on entering the terminal
  state. This necessarily means closing the GGPO session, as it's
  unsound to keep the session open _without_ also supporting simulation.
  (It can be done, it's just unsound.)
* Require players wait for spectators to enter the terminal state before
  terminating. This allows us to either close the GGPO session
  immediately or to preserve the GGPO session until spectators have
  terminated.

Closing the GGPO session immediately has some non-obvious consequences,
chief of which is that the spectator must return to the main menu after
the GGPO session is closed on the player side. If that's the case, it's
possible for a match on the spectator client to be shorter than a match
on the player client. That in itself isn't bad, but it has a knock-on
effect on state snapshots: it's possible for players to execute a frame
and capture a state snapshot from that frame, while the spectators never
will. This means the player clients may have permanently outstanding
state snapshots. We could work around this as well by just clearing the
state snapshots on terminal state, but that's a workaround on top of a
workaround on top of unsound behavior.

Instead, we should require player clients wait for termination, and
delay termination of the GGPO session until everyone has terminated.
This diff does not directly implement that, but it is a step towards
it: we now have arbitrary control over the terminal state and the
transition back to the main menu.