Sounds (particularly under GGPO): Implement "pure" sound management
Fixes #4.

SF4 natively attempts to play sounds directly inside the simulation.
Since that includes simulation steps executed during post-rollback
resimulation, rollbacks can result in heavily duplicated sounds. To
compound the problem further, the engine checks hardware state to
determine if a sound has stopped playing before returning the sound
handle to the object pool for reuse, sound overflow does _not_ cancel
older sounds (it stops the new sound from playing), and sounds are
stopped by handle (not by a cue ID).

We can solve most of those points by deferring playing sounds
until a sync step during predictive simulation, assuming all
sounds are permanently playing until imperatively stopped by the
engine, and changing the overflow strategy, but solving these in a
way that allows for the engine's existing handle-based stop and
mutation functions to work and to be totally deterministic
requires something (arguably too) clever.

The key is to create an additional SoundPlayerManager for every one
created by the vanilla engine. The one created by the engine is "pure",
and gets null player handles that don't ever actually operate on
anything, while the "shadow" one gets live player handles connected
to the sound playback library. When a sound is played on the "pure"
manager, track all the metadata about the sound, including which
player handle in the manager the sound is associated with. When saving
the game state, capture that metadata as well as all the player handles
in the manager's object pool. If we didn't save the player handles
as well as the metadata, the player handles would not remain stable
over the lifetime of a sound. However, the "shadow" manager with the
live player handles never gets its state saved or loaded- it's only
used as a way to imperatively play or stop sounds. Since the "shadow"
manager has the same sound metadata tracking as the "pure" manager,
synchronizing them involves stopping sounds in the "shadow" manager
that aren't playing the "pure" manager and vice versa. There are
additional tricks in tracking and synchronization that are way outside
the scope of this already-long commit message- if you find yourself
editing this code, vaya con dios.

In spirit, the approach is similar to Rollcaster's "How I didn’t
clip the sound effects" algorithm, but with a pretty radically
different implementation due to SF4's sound management and constraints
in the public release of GGPO. Where Rollcaster keeps a counter
of playing sounds and can freely consult previous states, sf4e
needs to respect the game's existing sound object pooling and GGPO
doesn't surface previous states. sf4e could certainly get closer
to Rollcaster's implementation, but I suspect it'd require a lot
more code.