Launcher: Wait for SF4 to be initialized before terminating
While this doesn't practically matter under Windows, there seems to be a
race condition on Linux when sf4e is run via Steam or protontricks and
the launcher closes before SF4 can start. When the launcher closes first,
SF4's `SteamAPI_Init()` call fails claiming the Steam services aren't
running. It's largely an ungrounded guess, but I can only assume this
has something to do with Steam's containerized runtime terminating
supporting services when the launched process exits, which means despite
the fact that SF4 runs in the containerized runtime there's no Steam
service to contact.

Instead, the launcher now waits for SF4 to be started, injects a
payload to the launched process, and waits on a Windows event. The sidecar
finds the payload, which includes a copy of the Windows event handle, and
sets the event during initialization. Observably, inserting a short sleep
in the launcher after launching the process would resolve the issue, but
I'd rather not introduce potential timing problems when fixing a timing
problem.