Dependency management: fix vcpkg baseline to actually contain versions
The baseline specified was actually too old- it didn't have the version
of imgui specified in the `vcpkg.json` manifest, resulting in sf4e
failing to get dependencies when run with the version of vcpkg specified
in the manifest. As always, it built on my machine anyway, thanks to the
version of vcpkg that I had _bootstrapped_ being different than the one
I had checked out.